Added my HalfwayFinishedStructureAnimation component
